Fitment for Linux and MacOs Build

+++ DigitalTwinLib/src/Model/Entities/Component.h
@@ -45,7 +45,7 @@ namespace DigitalTwin::Model {
         Variable* getMeasurable(std::string name);
         Variable* getAttribute(std::string name);
         Variable* resolveVariable(std::string name) override;
-        Variable* resolveVariable(std::vector<std::string> domains, int index) override;
+        Variable* resolveVariable(std::vector<std::string> domains, size_t index) override;
         Component* getComponent(std::string name);
         Port* getPort(std::string name);
